Fin servidor

+++ src/main/java/edu/escuealing/arep/calculadora/MergeSort.java
@@ -10,13 +10,13 @@ package edu.escuealing.arep.calculadora;
  * @author juan.navarro
  */
 public class MergeSort {
-    public static void mergeSort(Double[] a, int n) {
+    public static void mergeSort(int[] a, int n) {
         if (n < 2) {
             return;
         }
         int mid = n / 2;
-        Double[] l = new Double[mid];
-        Double[] r = new Double[n - mid];
+        int[] l = new int[mid];
+        int[] r = new int[n - mid];
 
         for (int i = 0; i < mid; i++) {
             l[i] = a[i];
@@ -30,7 +30,7 @@ public class MergeSort {
         merge(a, l, r, mid, n - mid);
     }
 
-    public static void merge(Double[] a, Double[] l, Double[] r, int left, int right) {
+    public static void merge(int[] a, int[] l, int[] r, int left, int right) {
 
         int i = 0, j = 0, k = 0;
         while (i < left && j < right) {
